Extract out with_env_tz helper method.
It’s used at so many places that extracting it out into a helper file
is worth doing.
